Closes #255

+++ src/main/java/biblivre/digitalmedia/DigitalMediaDAO.java
@@ -1,22 +1,3 @@
-/*******************************************************************************
- * Este arquivo é parte do Biblivre5.
- *
- * Biblivre5 é um software livre; você pode redistribuí-lo e/ou
- * modificá-lo dentro dos termos da Licença Pública Geral GNU como
- * publicada pela Fundação do Software Livre (FSF); na versão 3 da
- * Licença, ou (caso queira) qualquer versão posterior.
- *
- * Este programa é distribuído na esperança de que possa ser  útil,
- * mas SEM NENHUMA GARANTIA; nem mesmo a garantia implícita de
- * MERCANTIBILIDADE OU ADEQUAÇÃO PARA UM FIM PARTICULAR. Veja a
- * Licença Pública Geral GNU para maiores detalhes.
- *
- * Você deve ter recebido uma cópia da Licença Pública Geral GNU junto
- * com este programa, Se não, veja em <http://www.gnu.org/licenses/>.
- *
- * @author Alberto Wagner <alberto@biblivre.org.br>
- * @author Danniel Willian <danniel@biblivre.org.br>
- ******************************************************************************/
 package biblivre.digitalmedia;
 
 import java.io.File;
@@ -30,33 +11,30 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.postgresql.PGConnection;
-import org.postgresql.largeobject.LargeObject;
 import org.postgresql.largeobject.LargeObjectManager;
 
 import biblivre.core.AbstractDAO;
 import biblivre.core.exceptions.DAOException;
-import biblivre.core.file.DatabaseFile;
+import biblivre.core.file.BiblivreFile;
 import biblivre.core.file.MemoryFile;
 
-public class DigitalMediaDAO extends AbstractDAO {
+public abstract class DigitalMediaDAO extends AbstractDAO {
 
 	public static DigitalMediaDAO getInstance(String schema) {
-		return (DigitalMediaDAO) AbstractDAO.getInstance(DigitalMediaDAO.class, schema);
+		return (DigitalMediaDAO)
+			DigitalMediaDAOFactory.getDigitalMediaDAOImpl(schema);
+	}
+
+	protected abstract BiblivreFile getFile(long oid) throws Exception;
+
+	protected abstract void persist(InputStream is, long oid, long size) throws Exception;
+
+	protected DigitalMediaDAO() {
+		super();
 	}
 
 	public final Integer save(MemoryFile file) {
-		Connection con = null;
-
 		try (InputStream is = file.getNewInputStream()) {
-			con = this.getConnection();
-			con.setAutoCommit(false);
-
-			PGConnection pgcon = this.getPGConnection(con);
-
-			if (pgcon == null) {
-				throw new Exception("Invalid Delegating Connection");
-			}
-
 			Integer serial = file.getId();
 			if (serial == null) {
 				serial = this.getNextSerial("digital_media_id_seq");
@@ -64,73 +42,42 @@ public class DigitalMediaDAO extends AbstractDAO {
 			}
 
 			if (serial != 0) {
-				LargeObjectManager lobj = pgcon.getLargeObjectAPI();
-				long oid = lobj.createLO();
+				long oid = createOID();
 
-				LargeObject obj = lobj.open(oid, LargeObjectManager.WRITE);
+				persist(is, oid, file.getSize());
 
-				byte buf[] = new byte[4096];
-				int bytesRead = 0;
-				while ((bytesRead = is.read(buf)) > 0) {
-					obj.write(buf, 0, bytesRead);
+				try (Connection con2 = this.getConnection()) {
+					String sql = "INSERT INTO digital_media (id, name, blob, content_type, size) VALUES (?, ?, ?, ?, ?);";
+
+					PreparedStatement pst = con2.prepareStatement(sql);
+					pst.setInt(1, serial);
+					pst.setString(2, file.getName());
+					pst.setLong(3, oid);
+					pst.setString(4, file.getContentType());
+					pst.setLong(5, file.getSize());
+
+					pst.executeUpdate();
+					pst.close();
+					file.close();
 				}
-
-				obj.close();
-
-				String sql = "INSERT INTO digital_media (id, name, blob, content_type, size) VALUES (?, ?, ?, ?, ?);";
-
-				PreparedStatement pst = con.prepareStatement(sql);
-				pst.setInt(1, serial);
-				pst.setString(2, file.getName());
-				pst.setLong(3, oid);
-				pst.setString(4, file.getContentType());
-				pst.setLong(5, file.getSize());
-
-				pst.executeUpdate();
-				pst.close();
-				file.close();
-
-				this.commit(con);
-			} else {
-				this.rollback(con);
 			}
+
 			return serial;
 		} catch (Exception e) {
-			this.rollback(con);
 			throw new DAOException(e);
-		} finally {
-			this.closeConnection(con);
 		}
 	}
 
-
-	public final long importFile(File file) {
+	public long createOID() {
 		Connection con = null;
-
-		try (
-				InputStream is = new FileInputStream(file)
-				) {
+		try {
 			con = this.getConnection();
 			con.setAutoCommit(false);
 
-			PGConnection pgcon = this.getPGConnection(con);
-
-			if (pgcon == null) {
-				throw new Exception("Invalid Delegating Connection");
-			}
+			PGConnection pgcon = getPGConnection(con);
 
 			LargeObjectManager lobj = pgcon.getLargeObjectAPI();
 			long oid = lobj.createLO();
-
-			LargeObject obj = lobj.open(oid, LargeObjectManager.WRITE);
-
-			byte buf[] = new byte[4096];
-			int bytesRead = 0;
-			while ((bytesRead = is.read(buf)) > 0) {
-				obj.write(buf, 0, bytesRead);
-			}
-
-			obj.close();
 
 			this.commit(con);
 
@@ -143,22 +90,27 @@ public class DigitalMediaDAO extends AbstractDAO {
 		}
 	}
 
-	public final DatabaseFile load(int id, String name) {
+	public final long importFile(File file) {
 		Connection con = null;
-		DatabaseFile file = null;
 
-		try {
-			con = this.getConnection();
-			con.setAutoCommit(false);
+		try (InputStream is = new FileInputStream(file)) {
+			long oid = createOID();
 
-			PGConnection pgcon = this.getPGConnection(con);
+			persist(is, oid, file.length());
 
-			if (pgcon == null) {
-				throw new Exception("Invalid Delegating Connection");
-			}
+			return oid;
+		} catch (Exception e) {
+			this.rollback(con);
+			throw new DAOException(e);
+		} finally {
+			this.closeConnection(con);
+		}
+	}
 
-			LargeObjectManager lobj = pgcon.getLargeObjectAPI();
+	public final BiblivreFile load(int id, String name) {
+		BiblivreFile file = null;
 
+		try (Connection con = this.getConnection()) {
 			StringBuilder sql = new StringBuilder();
 			// We check both ID and FILE_NAME for security reasons, so users can't "guess"
 			// id's and get the files.
@@ -170,24 +122,21 @@ public class DigitalMediaDAO extends AbstractDAO {
 			pst.setString(2, name);
 
 			ResultSet rs = pst.executeQuery();
+
 			if (rs.next()) {
 				long oid = rs.getLong("blob");
-				LargeObject obj = lobj.open(oid, LargeObjectManager.READ);
 
-				file = new DatabaseFile(con, obj);
+				file = getFile(oid);
 
 				file.setName(rs.getString("name"));
 				file.setContentType(rs.getString("content_type"));
 				file.setLastModified(rs.getTimestamp("created").getTime());
 				file.setSize(rs.getLong("size"));
+
+				return file;
 			}
 		} catch (Exception e) {
-			this.rollback(con);
-			this.closeConnection(con);
-
 			throw new DAOException(e);
-		} finally {
-			// We must leave this connection open. file.close() will close it when needed.
 		}
 
 		return file;
@@ -197,24 +146,42 @@ public class DigitalMediaDAO extends AbstractDAO {
 		Connection con = null;
 		try {
 			con = this.getConnection();
-			StringBuilder sql = new StringBuilder();
-			// We check both ID and FILE_NAME for security reasons, so users can't "guess"
-			// id's and get the files.
-			sql.append("DELETE FROM digital_media ");
-			sql.append("WHERE id = ?;");
 
-			PreparedStatement pst = con.prepareStatement(sql.toString());
-			pst.setInt(1, id);
+			Statement st = con.createStatement();
+			ResultSet rs = st.executeQuery("SELECT id, blob, name FROM digital_media;");
 
-			int deleted = pst.executeUpdate();
-			// Find out if we need to check how many records were deleted from DB.
-			return deleted > 0;
+			if (rs.next()) {
+				DigitalMediaDTO dto = new DigitalMediaDTO();
+				dto.setId(rs.getInt("id"));
+				dto.setBlob(rs.getLong("blob"));
+				dto.setName(rs.getString("name"));
+
+				StringBuilder sql = new StringBuilder();
+				// We check both ID and FILE_NAME for security reasons, so users can't "guess"
+				// id's and get the files.
+				sql.append("DELETE FROM digital_media ");
+				sql.append("WHERE id = ?;");
+
+				PreparedStatement pst = con.prepareStatement(sql.toString());
+				pst.setInt(1, id);
+
+				int deleted = pst.executeUpdate();
+
+				deleteBlob(dto.getBlob());
+
+				// Find out if we need to check how many records were deleted from DB.
+				return deleted > 0;
+			}
+
+			return false;
 		} catch (Exception e) {
 			throw new DAOException(e);
 		} finally {
 			this.closeConnection(con);
 		}
 	}
+
+	protected abstract void deleteBlob(long blob);
 
 	public List<DigitalMediaDTO> list() {
 		Connection con = null;
@@ -239,5 +206,4 @@ public class DigitalMediaDAO extends AbstractDAO {
 			this.closeConnection(con);
 		}
 	}
-
 }
